Fix the DocFX documentation
