+++ Hazelcast.Net.sln
@@ -40,7 +40,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{25300DF5
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hazelcast.Net.Win32", "src\Hazelcast.Net.Win32\Hazelcast.Net.Win32.csproj", "{7E3B8518-73DB-471A-B624-B2FD6D8CCB61}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hazelcast.Net.DocAsCode", "src\Hazelcast.Net.DocAsCode\Hazelcast.Net.DocAsCode.csproj", "{FE505BEF-9319-4F93-A9D2-16CB721DBE44}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hazelcast.DocAsCode.Build", "src\Hazelcast.DocAsCode.Build\Hazelcast.DocAsCode.Build.csproj", "{FE505BEF-9319-4F93-A9D2-16CB721DBE44}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "doc", "doc", "{D974DFA3-D12C-4D97-9886-43A94FDA79D1}"
 	ProjectSection(SolutionItems) = preProject
