+++ doc/_docfx.json
@@ -11,7 +11,8 @@
       "dest": "obj/dev/api",
       "disableGitFeatures": true,
       "disableDefaultFilter": false,
-      "properties": { "TargetFramework": "netstandard2.0" }
+      "properties": { "TargetFramework": "net7.0" },
+      "memberLayout": "separatePages"
     }
   ],
   "build": {
@@ -34,7 +35,7 @@
     "overwrite": [ { "files": [ "apidoc/*.md" ] } ],
     "globalMetadata": {
       "_appTitle": "Hazelcast .NET Documentation",
-      "_appFooter": "Copyright &copy; 2010-2022 Hazelcast, Inc. All rights reserved.<br />Generated by DocFX.",
+      "_appFooter": "Copyright &copy; 2010-2023 Hazelcast, Inc. All rights reserved.<br />Generated by DocFX.",
       "_appFaviconPath": "images/favicon.ico",
       "_enableSearch": true,
       "_disableContribution": true
