+++ doc/dev/doc/configuration/options.md
@@ -4,6 +4,6 @@ comment: this file will be automatically complemented when building the doc
 ---
 # Configuration Options
 
-This page lists the complete set of options available to configure the Hazelcast .NET client.
-
-This page lists the complete set of options available to configure the Hazelcast .NET client. For more details regarding configuration in general, including configuration sources, refer to the [Configuration](../configuration.md) page.
+This page lists the complete set of options available to configure the Hazelcast .NET client. 
+For more details regarding configuration in general, including configuration sources,
+refer to the [Configuration](../configuration.md) page.
