+++ doc/templates/hz/partials/class.header.tmpl.partial
@@ -21,7 +21,7 @@
 </div>
 {{/inClass}}
 {{#implements.0}}
-<div classs="implements">
+<div class="implements">
   <h5>{{__global.implements}}</h5>
 {{/implements.0}}
 {{#implements}}
