+++ src/Hazelcast.DocAsCode.Build/Hazelcast.DocAsCode.Build.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <NoWarn>$(NoWarn);CS1591</NoWarn>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.DocAsCode.Common" Version="2.64.0" />
+    <PackageReference Include="Microsoft.DocAsCode.Build.Common" Version="2.64.0" />
+    <PackageReference Include="Microsoft.DocAsCode.Build.ConceptualDocuments" Version="2.64.0" />
+    <PackageReference Include="Microsoft.DocAsCode.Build.ManagedReference" Version="2.64.0" />
+    <PackageReference Include="Microsoft.DocAsCode.Build.TableOfContents" Version="2.64.0" />
+    <PackageReference Include="Microsoft.DocAsCode.Plugins" Version="2.64.0" />
+    <PackageReference Include="System.Composition" Version="7.0.0" />
+  </ItemGroup>
+
+</Project>
